[PyTorch] Decouple version numbers from c10 and caffe2 targets
There's size regression in model delivery in D25682312. Only the model version numbers are used. However, the dependency of the entire c10 (128 KB) is pulled in.

This diff is to decouple the version numbers to a separate header file, versions.h. Other targets referring to version numbers only can have deps of ```caffe2:version_headers```.

Differential Revision: [D25716601](https://our.internmc.facebook.com/intern/diff/D25716601/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D25716601/)!

ghstack-source-id: 119161467
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49905